requirements.txt
sarvamai
python-dotenv
flask 